Add file systems option and remove plugin-data
This updates the options to match changes in Chrome.

Plugin Data was deprecated many versions ago and so I am removing it. However, I still set the clear to remove it in case anyone is running a very old version, since it doesn't hurt anything.

I also added the "File systems" option, which appears to be a newer option for removal.