Switch user settings to be local
This changes the options from using sync to using local storage. The reason for this is so that users can have different settings on different devices.

However, I am leaving settings being saved in sync when they are changed. These synced settings are read on install and are used if no local options exist (local options have priority). This way, when a user installs the extension on a new device, the most recent settings will automatically be applied.